fix delete project

+++ pages/api/projects/[slug]/index.ts
@@ -1,5 +1,6 @@
 import { sendAudit } from '@/lib/retraced';
 import { getSession } from '@/lib/session';
+import { dropDatabase } from 'knex/knex.config';
 import {
   deleteProject,
   getProject,
@@ -108,14 +109,16 @@ const handleDELETE = async (req: NextApiRequest, res: NextApiResponse) => {
     });
   }
 
-  await deleteProject({ slug });
+  if (await dropDatabase(slug)) {
+    await deleteProject({ slug });
 
-  sendAudit({
-    action: 'project.delete',
-    crud: 'd',
-    user: session.user,
-    project,
-  });
+    sendAudit({
+      action: 'project.delete',
+      crud: 'd',
+      user: session.user,
+      project,
+    });
+  }
 
   return res.status(200).json({ data: {}, error: null });
 };
